moving card ascii tests over to test_card_ascii.py
